Extend fuzzer to full codec family; fix adaptive/FOR uninit-meta bug
What:
- varintPaletteFuzz grows to 9 strategies with a new direct round-trip
  oracle covering the previously-unfuzzed codecs: the adaptive selector,
  Elias gamma/delta arrays (>=1 domain), grouped varints (<=64 fields),
  and the u16 bitmap set (decode -> re-encode byte-identical).
- Abort-context instrumentation: a SIGABRT handler prints iteration,
  strategy slot, sub-phase, and stream count when a LIBRARY-side
  assert fires, making such failures as reproducible as harness checks.
  (Recorded triage lesson: a failure only in plain -O3 builds — passing
  at -O0 and under ASan/UBSan — usually means an uninitialized read,
  invisible to the available sanitizers.)
- Bug fix (found immediately by the new coverage): adaptive's FOR path
  passed an UNINITIALIZED varintFORMeta to varintFOREncode, whose
  intentional trust-the-caller fast path treats meta->count == count as
  "already analyzed". When -O3 stack reuse made the garbage match, FOR
  skipped analysis and encoded with a garbage offsetWidth (assert or
  corruption). Fixed at the callers — adaptive and two test sites now
  zero-init — preserving the zero-overhead API contract, which is now
  documented at varintFOREncode's declaration.
- Bug fix: varintDeltaDecodeUnsigned reconstructed with a signed add
  (UB on wrap); now unsigned, matching the encoder's wrapped subtract.
- README: 16 encodings, adds palette-of-deltas.

Why: compete only decodes the winning codec and the heuristic selector
was never exercised at all, so these paths had zero coverage; both new
bugs (7 total found by this fuzzer) were in exactly such paths.

How: validated with 18/18 ctest, 150k plain -O3 iterations on both
formerly-crashing seeds (2222, 88888) plus three control seeds, and a
400k-iteration ASan+UBSan session with zero reports.

+++ src/varintDelta.c
@@ -159,7 +159,9 @@ size_t varintDeltaDecodeUnsigned(const uint8_t *input, size_t count,
         varintWidth deltaBytes = varintDeltaGet(p, &delta);
         p += deltaBytes;
 
-        current = (uint64_t)((int64_t)current + delta);
+        /* Unsigned addition wraps mod 2^64, exactly reversing the
+         * encoder's wrapped subtraction; a signed add here is UB. */
+        current = current + (uint64_t)delta;
         output[i] = current;
     }
